Refactor code and update dependencies
Cleaned up code by deleting unused classes and files. Reworked the structure of the project moving files to a more appropriate package. Dependency on Mapstruct was introduced and adjusted configuration in pom.xml. Code has been refactored to create a clear separation of concerns, making the code easier to maintain and understand.