Remove test data and disable H2 database
The test data previously used for the application has been deleted and any dependency on the H2 database has been removed. Updates made are to the postgresql database configurations and for application production environment setup. The removed test data files and the disabling of H2 dependency improves application's production readiness.